ClusterIP from node support for Antrea Proxy on Linux
ClusterIP traffic from node will be routed to Antrea gateway.
The route on node is generated with ClusterIP. When a ClusterIP
Service is created, if there is no related route item,
AntreaProxy installs a route. If there is a route item
which can route the ClusterIP to Antrea gateway, do nothing.
If there is a route item and cannot route the ClusterIP to
Antrea gateway, extends the route IP scope with current
ClusterIP and installs a new route. The old route should be
also removed.

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>